+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 1.Fix bug </br>
 2.New organization fuction</br>
 3.New uploadImage method </br>
+4.Fix login slow bug</br>
 
 ## 2.0.8